Restructure lobby page files
