+++ src/components/LobbyPage/AliasModal.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import { db } from '../firebase';
-import { useUserId } from '../context/userContext';
+import { db } from '../../firebase';
+import { useUserId } from '../../context/userContext';
 import Modal from 'react-bootstrap/Modal';
 import Button from 'react-bootstrap/Button';
 import Form from 'react-bootstrap/Form';
