+++ src/components/LobbyPage/GameStart.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import { db } from '../firebase';
-import { useUserId } from '../context/userContext';
+import { db } from '../../firebase';
+import { useUserId } from '../../context/userContext';
 import Container from 'react-bootstrap/Container';
 import Button from 'react-bootstrap/Button';
 import Spinner from 'react-bootstrap/Spinner';
