+++ src/components/LobbyPage/index.js
@@ -1,0 +1,1 @@
+export { default } from './LobbyPage';
